pyproject.toml
[project]
name = "cutting-edge"
version = "0.1.0"
description = "AI-powered cutting stock optimization for sewing patterns"
authors = []
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "torch>=2.1.2",
    "torchvision>=0.16.2",
    "opencv-python>=4.9.0.80",
    "numpy>=1.26.3",
    "matplotlib>=3.8.2",
    "shapely>=2.0.7",
    "cairosvg>=2.7.1",
    "scikit-optimize>=0.9.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.285",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=cutting_edge"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = ["E501", "C901"]  # Line length, function complexity

[project.urls]
Homepage = "https://github.com/gdxbs/CuttingEdge"
Repository = "https://github.com/gdxbs/CuttingEdge.git"
Documentation = "https://github.com/gdxbs/CuttingEdge#readme"

[tool.hatch.build.targets.wheel]
packages = ["cutting_edge"]

[tool.hatch.metadata]
allow-direct-references = true
